vestiges + cleanup
